pyproject.toml
[project]
name = "i18n-lib"
version = "0.1.0"
description = "i18n lib"
readme = "README.md"
authors = [
    { name = "Hamlet", email = "hamlets849@gmail.com" }
]
license = "MIT"
keywords = ["i18n", "localization", "internationalization"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: Implementation :: CPython",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Internationalization",
    "Topic :: Software Development :: Localization",
    "Typing :: Typed"
]
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0.2",
]

[project.urls]
Homepage = "https://github.com/HamletSargsyan/i18n-lib"
Documentation = "https://hamletsargsyan.github.io/i18n-lib/usage"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "changelog-parser>=0.0.8",
    "mypy>=1.15.0",
    "pre-commit>=4.1.0",
    "pyright>=1.1.396",
    "pytest>=8.3.5",
    "ruff>=0.11.0",
    "semver>=3.0.4",
    "tomlkit>=0.13.2",
    "types-pyyaml>=6.0.12.20241230",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.9",
    "mkdocstrings[python]>=0.29.0",
]


[tool.hatch.build.targets.wheel]
packages = ["src/i18n"]

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true
